players/randomAI/aifleet/requirement.txt
django
wheel
django_rest_framework
django_redis
